move GreyBG to be next to BlueBG
