+++ src/pages/case-study/canon-giving-stories-a-new-form.js
@@ -10,7 +10,7 @@ import { SectionTitle, BodyPrimary } from '../../components/Typography'
 import CaseStudyHero from '../../components/Common/CaseStudyCards/CaseStudyHero'
 import CaseStudyPreview from '../../components/Common/CaseStudyCards/CaseStudyPreview'
 import Layout from '../../components/layout'
-import GreyBackground, { GreyBGOffset } from '../../components/GreyBG'
+import GreyBackground, { GreyBGOffset } from '../../components/Common/GreyBG'
 import landscape from '../../images/case-study/at_the_heart_of_a_story.svg'
 import Image from '../../components/Common/Image'
 import { makeText } from '../../utils/makeText'
