+++ src/pages/case-study/joyent-bringing-application-awareness-to-cloud.js
@@ -10,7 +10,7 @@ import { SectionTitle, BodyPrimary } from '../../components/Typography'
 import CaseStudyHero from '../../components/Common/CaseStudyCards/CaseStudyHero'
 import CaseStudyPreview from '../../components/Common/CaseStudyCards/CaseStudyPreview'
 import Layout from '../../components/layout'
-import GreyBackground from '../../components/GreyBG'
+import GreyBackground from '../../components/Common/GreyBG'
 import Image from '../../components/Common/Image'
 import { makeText } from '../../utils/makeText'
 import Head from '../../components/Common/Head'
