+++ src/templates/caseStudy.js
@@ -5,7 +5,7 @@ import { graphql } from 'gatsby'
 import generateCaseStudy from '../utils/generateCaseStudy'
 import Layout from '../components/layout'
 import { Grid } from '../components/grid'
-import { GreyBGOffset } from '../components/GreyBG'
+import { GreyBGOffset } from '../components/Common/GreyBG'
 import Head from '../components/Common/Head'
 import CaseStudyHero from '../components/Common/CaseStudyCards/CaseStudyHero'
 import FirstTextSection from '../components/TemplatedCaseStudy/FirstTextSection'
