+++ src/templates/service.js
@@ -10,7 +10,7 @@ import Layout from '../components/layout'
 import CaseStudyPreview from '../components/Common/CaseStudyCards/CaseStudyPreview'
 import SeoLinks from '../components/Common/seoLinks'
 import WorkStages from '../components/Service/WorkStages'
-import GreyBackground from '../components/GreyBG'
+import GreyBackground from '../components/Common/GreyBG'
 import BlueBackground from '../components/Common/BlueBG'
 import Head from '../components/Common/Head'
 import Statement from '../components/Common/Statement'
