+++ stories/titleAndList.stories.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf, addDecorator } from '@storybook/react'
 import Theme from './theme'
 import BlueBackground from '../src/components/Common/BlueBG'
-import GreyBackground from '../src/components/GreyBG'
+import GreyBackground from '../src/components/Common/GreyBG'
 import { Grid } from '../src/components/grid'
 import TitleAndList from '../src/components/Common/TitleAndList'
 import StyledLink from '../src/components/Common/StyledLink'
